# feature: md5'ing files without global lock