Directly md5 the files in C. This allows us to release the runtime lock
while md5'ing large files, which allows other OCaml threads to make
progress.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

<!-- ps-id: ad6d7706-eaf6-4e5a-b8a2-9669b8122db8 -->