@Alizter 2023-04-04
Bench results:
| | main | pr | %diff
| --- | --- | --- | --- |
| null | 0.8971710205078125 | 0.91907906532287598 | |
| | 0.91837000846862793 | 0.95876097679138184 | |
| | 0.90341901779174805 | 0.96033596992492676 | |
| | 0.8906099796295166 | 0.92078804969787598 | |
| | 0.89159893989562988 | 0.92145395278930664 | |
| avg. | 0.90023379325867 | 0.93608360290527 | +3.98228%  |
| clean | 107.84221076965332 | 107.52441096305847 | -0.29469% |