@rgrinberg 2023-04-04
Yeah, that's expected. The purpose of this PR is to stop hogging the OCaml lock to let other threads do work, but we aren't using threads for anything yet.