@rgrinberg 2023-04-04
>  Bench results:

I could not reproduce your results in #7489. The new digest is faster in fact. Which makes sense because this PR also the in_channel overhead.