@Alizter 2023-04-04
I'll try and do some more detailed benchmarks tomorrow.