@rgrinberg 2023-04-05
Here's the before (ignore the strings benchmarks):

```
_____________________________________________________________________________________________
_ Name                                       _     Time/Run _ mWd/Run _ mjWd/Run _ Prom/Run _ Percentage _
_____________________________________________________________________________________________
_ [bench/micro/digest_bench.ml] string:10    _     108.12ns _   4.00w _          _          _      0.02% _
_ [bench/micro/digest_bench.ml] string:100   _     187.23ns _   4.00w _          _          _      0.03% _
_ [bench/micro/digest_bench.ml] string:1000  _   1_391.95ns _   4.00w _          _          _      0.26% _
_ [bench/micro/digest_bench.ml] string:10000 _  13_389.54ns _   4.00w _          _          _      2.48% _
_ [bench/micro/digest_bench.ml] file:10      _ 311_854.72ns _   7.00w _    0.28w _    0.28w _     57.71% _
_ [bench/micro/digest_bench.ml] file:100     _ 319_334.90ns _   7.00w _    0.28w _    0.28w _     59.09% _
_ [bench/micro/digest_bench.ml] file:1000    _ 304_509.95ns _   7.00w _    0.28w _    0.28w _     56.35% _
_ [bench/micro/digest_bench.ml] file:10000   _ 321_887.22ns _   7.00w _    0.29w _    0.29w _     59.56% _
_ [bench/micro/digest_bench.ml] file:100000  _ 540_429.04ns _   7.00w _    0.30w _    0.30w _    100.00% _
_____________________________________________________________________________________________
```

And here's the after:

```
_____________________________________________________________________________________________
_ Name                                       _     Time/Run _ mWd/Run _ mjWd/Run _ Prom/Run _ Percentage _
_____________________________________________________________________________________________
_ [bench/micro/digest_bench.ml] string:10    _     117.81ns _   4.00w _          _          _      0.02% _
_ [bench/micro/digest_bench.ml] string:100   _     185.62ns _   4.00w _          _          _      0.04% _
_ [bench/micro/digest_bench.ml] string:1000  _   1_325.23ns _   4.00w _          _          _      0.28% _
_ [bench/micro/digest_bench.ml] string:10000 _  13_101.58ns _   4.00w _          _          _      2.74% _
_ [bench/micro/digest_bench.ml] file:10      _ 301_397.62ns _   7.00w _    0.28w _    0.28w _     63.11% _
_ [bench/micro/digest_bench.ml] file:100     _ 315_002.31ns _   7.00w _    0.28w _    0.28w _     65.95% _
_ [bench/micro/digest_bench.ml] file:1000    _ 306_201.32ns _   7.00w _    0.28w _    0.28w _     64.11% _
_ [bench/micro/digest_bench.ml] file:10000   _ 327_189.01ns _   7.00w _    0.29w _    0.29w _     68.51% _
_ [bench/micro/digest_bench.ml] file:100000  _ 477_604.25ns _   7.00w _    0.29w _    0.29w _    100.00% _
_____________________________________________________________________________________________
```

So we are a little faster (easiest to observe with the largest) file but the difference is mostly noise.

While the aim of this PR isn't to improve performance, I should note that it has the advantage of making dune itself more responsive. Currently, dune is unable to handle signals or file watching notifications while digesting large files. With this PR, it should be able to respond to these events even while digesting.