@rgrinberg 2023-04-05
The benchmark above is a bit messed up. I forgot to enable the new digest, and I was testing inside docker on mac which somewhat skewed the results. Here's more consistent numbers:

Before:

```
┌────────────────────────────────────────────┬──────────────┬─────────┬──────────┬──────────┬────────────┐
│ Name                                       │     Time/Run │ mWd/Run │ mjWd/Run │ Prom/Run │ Percentage │
├────────────────────────────────────────────┼──────────────┼─────────┼──────────┼──────────┼────────────┤
│ [bench/micro/digest_bench.ml] string:10    │     141.02ns │   4.00w │          │          │      0.07% │
│ [bench/micro/digest_bench.ml] string:100   │     245.25ns │   4.00w │          │          │      0.11% │
│ [bench/micro/digest_bench.ml] string:1000  │   1_891.66ns │   4.00w │          │          │      0.88% │
│ [bench/micro/digest_bench.ml] string:10000 │  17_588.78ns │   4.00w │          │          │      8.15% │
│ [bench/micro/digest_bench.ml] file:10      │  29_836.32ns │   7.00w │    0.27w │    0.27w │     13.82% │
│ [bench/micro/digest_bench.ml] file:100     │  27_164.67ns │   7.00w │    0.27w │    0.27w │     12.58% │
│ [bench/micro/digest_bench.ml] file:1000    │  30_303.51ns │   7.00w │    0.27w │    0.27w │     14.04% │
│ [bench/micro/digest_bench.ml] file:10000   │  45_642.59ns │   7.00w │    0.27w │    0.27w │     21.14% │
│ [bench/micro/digest_bench.ml] file:100000  │ 215_870.58ns │   7.00w │    0.28w │    0.28w │    100.00% │
└────────────────────────────────────────────┴──────────────┴─────────┴──────────┴──────────┴────────────┘
```

After:
```
┌────────────────────────────────────────────┬──────────────┬─────────┬────────────┐                                                                                                                                        │ Name                                       │     Time/Run │ mWd/Run │ Percentage │                                                                                                                                        ├────────────────────────────────────────────┼──────────────┼─────────┼────────────┤                                                                                                                                        │ [bench/micro/digest_bench.ml] string:10    │     158.99ns │   4.00w │      0.09% │                                                                                                                                        │ [bench/micro/digest_bench.ml] string:100   │     230.71ns │   4.00w │      0.13% │                                                                                                                                        │ [bench/micro/digest_bench.ml] string:1000  │   1_774.27ns │   4.00w │      0.99% │                                                                                                                                        │ [bench/micro/digest_bench.ml] string:10000 │  16_723.45ns │   4.00w │      9.34% │                                                                                                                                        │ [bench/micro/digest_bench.ml] file:10      │   6_611.19ns │   4.00w │      3.69% │                                                                                                                                        │ [bench/micro/digest_bench.ml] file:100     │   6_682.69ns │   4.00w │      3.73% │
│ [bench/micro/digest_bench.ml] file:1000    │   8_393.22ns │   4.00w │      4.69% │
│ [bench/micro/digest_bench.ml] file:10000   │  23_870.62ns │   4.00w │     13.33% │
│ [bench/micro/digest_bench.ml] file:100000  │ 179_108.05ns │   4.00w │    100.00% │
└────────────────────────────────────────────┴──────────────┴─────────┴────────────┘
```

@Alizter try confirming this