@Alizter 2023-04-06
#### Main
```
Benchmark 1: ./dune.exe build @install --cache disabled
  Time (mean ± σ):     10.245 s ±  0.238 s    [User: 86.558 s, System: 30.658 s]
  Range (min … max):    9.672 s … 10.523 s    10 runs
```
#### PR
```
Benchmark 1: ./dune.exe build @install --cache disabled
  Time (mean ± σ):     10.160 s ±  0.112 s    [User: 85.362 s, System: 30.681 s]
  Range (min … max):    9.898 s … 10.261 s    10 runs
```

The reason I benched originally was to check for any potential slowdown. The builds of the Dune repo I did pretty much confirm that there is no slow down and maybe a little speed up.